refactor(plugin-server): merge e2e tests
